chore(scap): move scap_proc_fill_cgroups to scap_cgroup.c
