+++ userspace/libscap/linux/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_library(scap_platform scap_linux_platform.c scap_procs.c scap_fds.c scap_userlist.c scap_iflist.c scap_ppm_sc.c)
+add_library(scap_platform scap_linux_platform.c scap_procs.c scap_fds.c scap_userlist.c scap_iflist.c scap_ppm_sc.c scap_cgroup.c)
 target_link_libraries(scap_platform scap_error)
 set_scap_target_properties(scap_platform)
